Add compatibility for plugins providing their own SimpleCommandMap (Fixes #11)
This commit adds compatibility for plugins which replace Bukkit's SimpleCommandMap with their own version *and copy field data over*, such as PerWorldPlugins. It does so by traversing the class hierarchy and looking for a SimpleCommandMap. If it doesn't find any, a fancy error message is displayed with plugin and server info as well as some information about the unexpected class which has been encountered.